drive file upload bug fix

+++ app/Http/Controllers/Admin/CarDriverController.php
@@ -65,13 +65,12 @@ class CarDriverController extends BaseController
     {
         $model = CarDriver::find($id);
         $req = $request->all();
-        if($model->grid_image){
-            if(isset($req['grid_image']) && file_exists(public_path('uploads/'.$model->grid_image))) {
-                unlink(public_path() . '/uploads/' . $model->grid_image);
-                $req['grid_image'] = $this->fileUpload($request->grid_image, public_path('uploads/'))[0];
-            }else{
-                $req['grid_image'] = $this->fileUpload($request->grid_image, public_path('uploads/'))[0];
-            }
+
+        if($model->grid_image && isset($req['grid_image']) && file_exists(public_path('uploads/'.$model->grid_image))) {
+            unlink(public_path() . '/uploads/' . $model->grid_image);
+            $req['grid_image'] = $this->fileUpload($request->grid_image, public_path('uploads/'))[0];
+        }else{
+            $req['grid_image'] = $this->fileUpload($request->grid_image, public_path('uploads/'))[0];
         }
 
         $req['with_driver'] = isset($req['with_driver']);
@@ -79,7 +78,7 @@ class CarDriverController extends BaseController
         if($req['with_driver'] != $model->with_driver) {
             $model['data'] = '';
         }
-        dd($req);
+
         $model->update($req);
         $images = $this->fileUpload($request->file('image'), public_path('uploads/'));
